% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{convert_to_rstan}
\alias{convert_to_rstan}
\title{Convert NumPyro posterior to rstan-like format.}
\usage{
convert_to_rstan(fit)
}
\arguments{
\item{fit}{A mrhevo_numpyro object.}
}
\value{
A list with posterior samples in rstan-compatible format.
}
\description{
Convert NumPyro posterior to rstan-like format.
}
