% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{format.z.aspvalue}
\alias{format.z.aspvalue}
\title{Generate formatted p-values from z values.}
\usage{
\method{format}{z.aspvalue}(z, sigfig = 1, neglogp.threshold.scinot = 3, neglogp.threshold = NULL)
}
\arguments{
\item{z}{Standard normal deviate.}

\item{sigfig}{Number of significant figures to show in p-values.}

\item{neglogp.threshold.scinot}{Minus log10(pvalue) threshold for scientific
notation.}

\item{neglogp.threshold}{Minus log10(pvalue) cutoff for thresholding
p-values.}
}
\value{
Vector of formatted p-values.
}
\description{
Generate formatted p-values from z values.
}
