% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{format_scinot_pvalue}
\alias{format_scinot_pvalue}
\title{Reformat p-values that are in scientific notation for LaTeX.}
\usage{
format_scinot_pvalue(x)
}
\arguments{
\item{x}{Character vector of numbers in scientific notation.}
}
\value{
A character vector of LaTeX math expressions.
}
\description{
Reformat p-values that are in scientific notation for LaTeX.
}
