% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{get_coeffratios}
\alias{get_coeffratios}
\title{calculate ratios of coefficients}
\usage{
get_coeffratios(coeffs.dt, use.delta = FALSE)
}
\value{
Input data.table plus a column containing ratios of coefficients
              for each genetic instrument.
}
\description{
calculate ratios of coefficients
}
