% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{get_estimatorsMR}
\alias{get_estimatorsMR}
\title{Calculate inverse-variance weighted MR estimator from summary stats.}
\usage{
get_estimatorsMR(coeffs.dt)
}
\arguments{
\item{coeffs.dt}{Data.table with columns for coefficient and SE of
regression of exposure on instrument: \code{alpha_hat},
\code{se.alpha_hat}),
regression of outcome on instrument: \code{gamma_hat},
\code{se.gamma_hat} and coefficient ratios:
\code{theta_IV}, \code{se.theta_IV}.}
}
\value{
A data.table with inverse-variance weighted estimate.
}
\description{
Calculate inverse-variance weighted MR estimator from summary stats.
}
