% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{get_summarystatsforMR}
\alias{get_summarystatsforMR}
\title{Calculate summary stats for second step of two-step Mendelian randomization.}
\usage{
get_summarystatsforMR(Y, Z, X_u)
}
\arguments{
\item{Y}{Integer vector of binary outcome or numeric vector of continuous
outcome.}

\item{Z}{Data.table of genetic instruments.}

\item{X_u}{Data.table of covariates.}
}
\value{
Data.table of coefficients for each genetic instrument
}
\description{
Calculate summary stats for second step of two-step Mendelian randomization.
}
