% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{mle.se.pval}
\alias{mle.se.pval}
\title{Calculate maximum likelihood estimate and p-value from posterior samples and
prior.}
\usage{
mle.se.pval(x, prior, return.asplot = FALSE)
}
\arguments{
\item{x}{Numeric vector of samples from the posterior distribution of the
parameter.}

\item{prior}{Numeric vector of values of the prior density for each element
of x.}

\item{return.asplot}{Logical value determines whether the function returns
maximum likelihood estimate or a plot of the posterior
density and log-likelihood.}
}
\value{
If \code{return.asplot} is FALSE, return data.table with one row
containing maximum likelihood estimate, standard error,
test statistic, p-value, formatted p-value. Otherwise, create a plot
of posterior density and log-likelihood.
}
\description{
Calculate maximum likelihood estimate and p-value from posterior samples and
prior.
}
