% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{msg}
\alias{msg}
\title{Report a message to the terminal}
\usage{
msg(mode, ..., LF = TRUE)
}
\arguments{
\item{mode}{One of \sQuote{info} (normal messages), \sQuote{note} (messages
that require some highlighting), \sQuote{warn} (important information
the user should definitely notice).}

\item{...}{Strings to be reported.}

\item{LF}{Whether a newline character should be added at the end of the
message (\code{TRUE} by default).}
}
\description{
Report a message to the terminal
}
