% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{plot_iv_estimates}
\alias{plot_iv_estimates}
\title{Plot instrumental variable estimates with MLE as slope line.}
\usage{
plot_iv_estimates(
  alpha_hat,
  se.alpha_hat,
  gamma_hat,
  se.gamma_hat,
  theta,
  qtlname = NULL
)
}
\arguments{
\item{alpha_hat}{Vector of estimated coefficients for effect of instruments
on exposure.}

\item{se.alpha_hat}{Vector of standard errors for coefficients alpha_hat.}

\item{gamma_hat}{Vector of estimated coefficients for effect of instruments
on outcome.}

\item{se.gamma_hat}{Vector of standard errors for coefficients gamma_hat.}

\item{theta}{Maximum likelihood estimate of causal effect (slope).}

\item{qtlname}{Optional vector of QTL names for labeling points.}
}
\value{
A ggplot object showing IV estimates with MLE as line through origin.
}
\description{
Plot instrumental variable estimates with MLE as slope line.
}
