% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{plot_kappa_hist}
\alias{plot_kappa_hist}
\title{Plot histogram of kappa shrinkage coefficients.}
\usage{
plot_kappa_hist(fit, bin_width = 0.02)
}
\arguments{
\item{fit}{Stan fit object or mrhevo_numpyro object.}

\item{bin_width}{Width of bins for histogram (default 0.02).}
}
\value{
A ggplot object showing histogram of kappa values.
}
\description{
Plot histogram of kappa shrinkage coefficients.
}
