% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{plot_posterior_pairs}
\alias{plot_posterior_pairs}
\title{Plot pairs of posterior samples.}
\usage{
plot_posterior_pairs(fit, pars = c("log_tau", "log_eta", "f"))
}
\arguments{
\item{fit}{Stan fit object or mrhevo_numpyro object.}

\item{pars}{Vector of parameter names to plot (default c("log_tau", "log_eta", "f")).}
}
\value{
A ggplot object showing pairs plot.
}
\description{
Plot pairs of posterior samples.
}
