% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{run_mrhevo.fixedtau}
\alias{run_mrhevo.fixedtau}
\title{Run Stan model for Mendelian randomization with regularized horsehoe prior
on pleiotropic effects, using summary statistics only.}
\usage{
run_mrhevo.fixedtau(
  alpha_hat,
  se.alpha_hat,
  gamma_hat,
  se.gamma_hat,
  tau = 1e-06,
  slab_scale = 0.2,
  slab_df = 2,
  priorsd_theta = 1,
  model.dir = system.file("stan", package = "mrhevo")
)
}
\arguments{
\item{alpha_hat}{Vector of estimated coefficients for effect of instruments
on exposure.}

\item{se.alpha_hat}{Vector of standard errors for coefficients alpha_hat.}

\item{gamma_hat}{Vector of estimated coefficients for effect of instruments
on outcome.}

\item{se.gamma_hat}{Vector of standard errors for coefficients gamma_hat.}

\item{tau}{Fixed value for global shrinkage parameter (default 1e-6).}

\item{slab_scale}{Scale parameter for slab component of regularized horseshoe
prior (default 0.2).}

\item{slab_df}{Degrees of freedom for slab component of regularized horseshoe
prior (default 2).}

\item{priorsd_theta}{Standard deviation of prior on theta.}

\item{model.dir}{Full path to STAN model directory
(default: bundled Stan models).}
}
\value{
An object of class stanfit.
}
\description{
Run Stan model for Mendelian randomization with regularized horsehoe prior
on pleiotropic effects, using summary statistics only.
}
