% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{format.scinot.pvalue}
\alias{format.scinot.pvalue}
\title{reformat for LaTeX p-values that are in scientific notation}
\usage{
\method{format}{scinot.pvalue}(x)
}
\arguments{
\item{x}{Character vector of numbers in scientific notation.}
}
\value{
A character vector of LaTeX math expressions.
}
\description{
reformat for LaTeX p-values that are in scientific notation
}
