% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{get_coeffratios}
\alias{get_coeffratios}
\title{Calculate ratios of coefficients.}
\usage{
get_coeffratios(coeffs.dt, use.delta = FALSE)
}
\arguments{
\item{coeffs.dt}{Data table with coefficients for each genetic instrument.}
}
\value{
A data.table with coefficients for genetic instruments plus a column
containing ratios of coefficients for each genetic instrument.
}
\description{
Calculate ratios of coefficients.
}
