% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{get_estimatorsMR}
\alias{get_estimatorsMR}
\title{Calculate conventional MR estimators from summary stats.}
\usage{
get_estimatorsMR(coeffs.dt)
}
\arguments{
\item{coeffs.dt}{Data.table with columns for coefficient and SE of
regression of exposure on instrument: \code{alpha_hat},
\code{se.alpha_hat}),
regression of outcome on instrument: \code{gamma_hat},
\code{se.gamma_hat} and coefficient ratios:
\code{theta_IV}, \code{se.theta_IV}.}
}
\value{
A data.table of estimates for weighted mean, weighted median, and
penalized weighted median of instrumental variable estimates.
}
\description{
Calculate conventional MR estimators from summary stats.
}
\note{
Details on penalized weighted median estimator can be found in
Bowden J, Davey Smith G, Haycock PC, Burgess S. Consistent Estimation
in Mendelian Randomization with Some Invalid Instruments Using a
Weighted Median Estimator. Genet Epidemiol. 2016 May;40(4):304-14.
doi: 10.1002/gepi.21965. Epub 2016 Apr 7. PMID: 27061298;
PMCID: PMC4849733.
}
