% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{get_estimatorsMR}
\alias{get_estimatorsMR}
\title{calculate conventional MR estimators from summary stats}
\usage{
get_estimatorsMR(coeffs.dt)
}
\arguments{
\item{coeffs.dt}{Data.table with columns for coefficient and SE of regression of exposure on instrument (alpha_hat, se.alpha_hat), regression of outcome on instrument (gamma_hat, se.gamma_hat) and coefficient ratios (theta_IV, se.theta_IV).}
}
\value{
Table of estimates for weighted mean, weighted median, and penalized weighted median of instrumental variable estimates.
}
\description{
calculate conventional MR estimators from summary stats
}
