% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{mle.se.pval}
\alias{mle.se.pval}
\title{calculate maximum likelihood estimate and p-value from posterior samples and prior}
\usage{
mle.se.pval(x, prior, return.asplot = FALSE)
}
\arguments{
\item{x}{Numeric vector of samples from the posterior distribution of the parameter.}

\item{prior}{Numeric vector of values of the prior density for each element of x.}

\item{return.asplot}{Logical value determines whether the function returns maximum likelihood estimate or a plot of the posterior density and log-likelihood.}
}
\value{
If return.asplot is FALSE, data.table with one row containing maximum likelihood estimate, standard error, test statistic, p-value, formatted p-value.
}
\description{
calculate maximum likelihood estimate and p-value from posterior samples and prior
}
