% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrhevo.R
\docType{package}
\name{mrhevo-package}
\alias{mrhevo}
\alias{mrhevo-package}
\title{Inference of causal effects by Mendelian randomisation, marginalising over
distribution of direct effects}
\description{
Inference of causal effects by Mendelian randomization, marginalizing over distribution of direct effects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/molepi-precmed/mrhevo}
}

}
\author{
Paul McKeigue \email{paul.mckeigue@ed.ac.uk},
Andrii Iakovliev \email{andrii.iakovliev@ed.ac.uk},
Buddhi Erabadda \email{b.erabadda@ed.ac.uk},
Athina Spiliopoulou \email{a.spiliopoulou@ed.ac.uk}
}
