% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{pnorm.extreme}
\alias{pnorm.extreme}
\title{Return upper bound on p-value where z is so extreme that pnorm(-z) returns 0.}
\usage{
pnorm.extreme(z, upper = TRUE)
}
\arguments{
\item{z}{Standard normal deviate.}
}
\value{
A string in scientific notation.
}
\description{
Return upper bound on p-value where z is so extreme that pnorm(-z) returns 0.
}
