% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{run_mrhevo.fixedtau}
\alias{run_mrhevo.fixedtau}
\title{Run Stan model for Mendelian randomization with regularized horseshoe prior
on pleiotropic effects, using summary statistics only.}
\usage{
run_mrhevo.fixedtau(
  alpha_hat,
  se.alpha_hat,
  gamma_hat,
  se.gamma_hat,
  tau = 1e-06,
  slab_scale = 0.2,
  slab_df = 2,
  priorsd_theta = 1,
  model.dir
)
}
\arguments{
\item{alpha_hat}{Vector of estimated coefficients for effect of instruments
on exposure.}

\item{se.alpha_hat}{Vector of standard errors for coefficients alpha_hat.}

\item{gamma_hat}{Vector of estimated coefficients for effect of instruments
on outcome.}

\item{se.gamma_hat}{Vector of standard errors for coefficients gamma_hat.}

\item{slab_scale}{scale param of prior on direct effects.}

\item{priorsd_theta}{Standard deviation of prior on theta.}

\item{model.dir}{Full path to STAN model directory.}

\item{fraction_pleio}{Prior guess at fraction of instruments that have
pleiotropic effects: values between 0.05 and 0.95 are allowed.}
}
\value{
An object of class stanfit.
}
\description{
Run Stan model for Mendelian randomization with regularized horseshoe prior
on pleiotropic effects, using summary statistics only.
}
