% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.mrhevo.R
\name{set.tau0}
\alias{set.tau0}
\title{Get value \code{tau0} for global shrinkage parameter \code{tau} given
expectation of \code{fraction_pleio}.}
\usage{
set.tau0(fraction_pleio = NULL, nu_global = 1, J, info)
}
\arguments{
\item{fraction_pleio}{Prior guess at fraction of effects that are nonzero.}

\item{nu_global}{Shape parameter of gamma prior.}

\item{J}{number of variables.}

\item{pseudovariance}{Variance of outcome variable.}
}
\value{
Value of tau0.
}
\description{
Get value \code{tau0} for global shrinkage parameter \code{tau} given
expectation of \code{fraction_pleio}.
}
